ML/app/requirements.txt
flask
flask-cors
scikit-learn
tensorflow
numpy
Pillow
